fmt: Fixing project structure
Re-aligning the project to be more organized like snpguest.
